Security patch for not letting users to access other user's data.
Change : Added Authentication authentication to getAllTask api

With this new change,
a. let's say if user hits getAllTask api then the api will return all the tasks which belong to the logged in user

b. let's say if user hits getAllTask with tasklistId then the api will return all the tasks from the provided tasklist id only if those tasks belongs to the logged in user.